made biochatter an agent
