--- examples/bio_ai/bioatomspace_data_subset/reactome/nodes.metta
@@ -1,58 +1,0 @@
-(pathway R-HSA-9662851)
-(name (pathway R-HSA-9662851) Anti-inflammatory_response_favouring_Leishmania_parasite_infection)
-(pathway R-HSA-140875)
-(name (pathway R-HSA-140875) Common_Pathway_of_Fibrin_Clot_Formation)
-(pathway R-HSA-1280215)
-(name (pathway R-HSA-1280215) Cytokine_Signaling_in_Immune_system)
-(pathway R-HSA-1266738)
-(name (pathway R-HSA-1266738) Developmental_Biology)
-(pathway R-HSA-1643685)
-(name (pathway R-HSA-1643685) Disease)
-(pathway R-HSA-2029481)
-(name (pathway R-HSA-2029481) FCGR_activation)
-(pathway R-HSA-9664323)
-(name (pathway R-HSA-9664323) FCGR3A-mediated_IL10_synthesis)
-(pathway R-HSA-9664422)
-(name (pathway R-HSA-9664422) FCGR3A-mediated_phagocytosis)
-(pathway R-HSA-2029480)
-(name (pathway R-HSA-2029480) Fcgamma_receptor_\(FCGR\)_dependent_phagocytosis)
-(pathway R-HSA-140877)
-(name (pathway R-HSA-140877) Formation_of_Fibrin_Clot_\(Clotting_Cascade\))
-(pathway R-HSA-6809371)
-(name (pathway R-HSA-6809371) Formation_of_the_cornified_envelope)
-(pathway R-HSA-109582)
-(name (pathway R-HSA-109582) Hemostasis)
-(pathway R-HSA-168256)
-(name (pathway R-HSA-168256) Immune_System)
-(pathway R-HSA-5663205)
-(name (pathway R-HSA-5663205) Infectious_disease)
-(pathway R-HSA-168249)
-(name (pathway R-HSA-168249) Innate_Immune_System)
-(pathway R-HSA-6785807)
-(name (pathway R-HSA-6785807) Interleukin-4_and_Interleukin-13_signaling)
-(pathway R-HSA-6805567)
-(name (pathway R-HSA-6805567) Keratinization)
-(pathway R-HSA-9658195)
-(name (pathway R-HSA-9658195) Leishmania_infection)
-(pathway R-HSA-9664433)
-(name (pathway R-HSA-9664433) Leishmania_parasite_growth_and_survival)
-(pathway R-HSA-9664417)
-(name (pathway R-HSA-9664417) Leishmania_phagocytosis)
-(pathway R-HSA-6798695)
-(name (pathway R-HSA-6798695) Neutrophil_degranulation)
-(pathway R-HSA-9664407)
-(name (pathway R-HSA-9664407) Parasite_infection)
-(pathway R-HSA-9824443)
-(name (pathway R-HSA-9824443) Parasitic_Infection_Pathways)
-(pathway R-HSA-76002)
-(name (pathway R-HSA-76002) Platelet_activation,_signaling_and_aggregation)
-(pathway R-HSA-114608)
-(name (pathway R-HSA-114608) Platelet_degranulation_)
-(pathway R-HSA-418346)
-(name (pathway R-HSA-418346) Platelet_homeostasis)
-(pathway R-HSA-432142)
-(name (pathway R-HSA-432142) Platelet_sensitization_by_LDL)
-(pathway R-HSA-76005)
-(name (pathway R-HSA-76005) Response_to_elevated_platelet_cytosolic_Ca2+)
-(pathway R-HSA-449147)
-(name (pathway R-HSA-449147) Signaling_by_Interleukins)
