--- examples/bio_ai/bioatomspace_data_subset/uniprot/edges.metta
@@ -1,16 +1,0 @@
-(translates_to (transcript ENST00000374003) (protein P09769))
-(translates_to (transcript ENST00000374004) (protein P09769))
-(translates_to (transcript ENST00000374005) (protein P09769))
-(translates_to (transcript ENST00000399173) (protein P09769))
-(translates_to (transcript ENST00000335071) (protein Q13461))
-(translates_to (transcript ENST00000256084) (protein Q9NQ38))
-(translates_to (transcript ENST00000359874) (protein Q9NQ38))
-(translates_to (transcript ENST00000398454) (protein Q9NQ38))
-(translation_of (protein P09769) (transcript ENST00000374003))
-(translation_of (protein P09769) (transcript ENST00000374004))
-(translation_of (protein P09769) (transcript ENST00000374005))
-(translation_of (protein P09769) (transcript ENST00000399173))
-(translation_of (protein Q13461) (transcript ENST00000335071))
-(translation_of (protein Q9NQ38) (transcript ENST00000256084))
-(translation_of (protein Q9NQ38) (transcript ENST00000359874))
-(translation_of (protein Q9NQ38) (transcript ENST00000398454))
