--- examples/bio_ai/biochatter-metta/.gitignore
@@ -1,21 +1,0 @@
-.DS_Store
-.vscode
-.github
-dist/
-__pycache__/
-.venv
-venv
-my_test/
-.pytest_cache
-.env
-*.mp3
-.cache
-*.env.idea
-.idea/
-*.env
-volumes/
-benchmark/encrypted_llm_test_data.json
-site/
-benchmark/data/*
-!benchmark/data/benchmark_data.json
-serve.sh
