--- examples/bio_ai/biochatter-metta/__init__.py
@@ -1,1 +1,0 @@
-from .main import biochatter_atom
