--- examples/bio_ai/biochatter-metta/pyproject.toml
@@ -1,35 +1,0 @@
-[tool.poetry]
-name = "biochatter-metta"
-version = "0.1.0"
-description = ""
-authors = ["Tibex88 <tibesolomon7@gmail.com>"]
-readme = "README.md"
-
-[tool.poetry.dependencies]
-python = ">=3.10,<3.12"
-hyperon = "0.1.6"
-langchain = "^0.0.347"
-openai = "^1.1.0"
-pymupdf = "^1.22.3"
-pydantic = "1.10.13"
-pymilvus = "2.2.8"
-tiktoken = "^0.5.2"
-nltk = "^3.8.1"
-redis = "^4.5.5"
-retry = "^0.9.2"
-stringcase = "^1.2.0"
-transformers = "^4.30.2"
-streamlit = { version = "^1.23.1", optional = true }
-gTTS = { version = "^2.3.2", optional = true }
-botocore = { version = "^1.33.9", optional = true }
-xinference = { version = "^0.8.4", optional = true }
-rsa = "^4.9"
-cryptography = "^41.0.7"
-neo4j-utils = "^0.0.7"
-seaborn = "^0.13.2"
-setuptools = "69.1.1"
-
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
