--- examples/bio_ai/biochatter-metta/requirements.txt
@@ -1,20 +1,0 @@
-hyperon==0.1.6
-langchain==0.0.347
-openai==1.1.0
-pymupdf==1.22.3
-pydantic==1.10.13
-pymilvus==2.2.8
-tiktoken==0.5.2
-nltk==3.8.1
-redis==4.5.5
-retry==0.9.2
-stringcase==1.2.0
-transformers==4.30.2
-# streamlit==1.23.1
-# gTTS==2.3.2
-# botocore==1.33.9
-# xinference==0.8.4
-rsa==4.9
-cryptography==41.0.7
-neo4j-utils==0.0.7
-seaborn==0.13.2
