+++ examples/bio_ai/motto/agents/biochatter/.gitignore
@@ -1,0 +1,21 @@
+.DS_Store
+.vscode
+.github
+dist/
+__pycache__/
+.venv
+venv
+my_test/
+.pytest_cache
+.env
+*.mp3
+.cache
+*.env.idea
+.idea/
+*.env
+volumes/
+benchmark/encrypted_llm_test_data.json
+site/
+benchmark/data/*
+!benchmark/data/benchmark_data.json
+serve.sh
