+++ examples/bio_ai/motto/agents/biochatter/biocypher_config/biocypher_config.yaml
@@ -1,0 +1,26 @@
+# add your settings here (overriding the defaults)
+
+biocypher:
+  offline: true
+  debug: false
+  schema_config_path: config/schema_config.yaml
+
+  # Ontology configuration
+  head_ontology:
+    url: https://github.com/biolink/biolink-model/raw/v3.2.1/biolink-model.owl.ttl
+    root_node: entity
+  
+
+  tail_ontologies:
+
+    cheminf:
+      url: http://purl.obolibrary.org/obo/cheminf.owl
+      head_join_node: chemical entity
+      tail_join_node: has_same_connectivity_as
+      merge_nodes: false
+    
+    cheminf_pubchem:
+      url: http://purl.obolibrary.org/obo/cheminf.owl
+      head_join_node: chemical entity 
+      tail_join_node: has PubChem normalized counterpart
+      merge_nodes: false
