+++ examples/bio_ai/requirements.py
@@ -11,3 +11,4 @@ pymupdf==1.22.3
 pymilvus==2.2.8
 transformers==4.30.2
 retry==0.9.2
+redis==4.5.5
